+++ aws_emr_launch/constructs/step_functions/emr_launch_function.py
@@ -42,7 +42,7 @@ class EMRLaunchFunctionNotFoundError(Exception):
 class EMRLaunchFunction(core.Construct):
     def __init__(self, scope: core.Construct, id: str, *,
                  launch_function_name: str,
-                 cluster_config: Optional[ClusterConfiguration] = None,
+                 cluster_configuration: Optional[ClusterConfiguration] = None,
                  namespace: str = 'default',
                  default_fail_if_cluster_running: bool = False,
                  success_topic: Optional[sns.Topic] = None,
@@ -52,9 +52,16 @@ class EMRLaunchFunction(core.Construct):
                  description: Optional[str] = None) -> None:
         super().__init__(scope, id)
 
-        if cluster_config is None:
+        if cluster_configuration is None:
             return
 
+        self._launch_function_name = launch_function_name
+        self._namespace = namespace
+        self._cluster_configuration = cluster_configuration
+        self._default_fail_if_cluster_running = default_fail_if_cluster_running
+        self._success_topic = success_topic
+        self._failure_topic = failure_topic
+        self._override_cluster_configs_lambda = override_cluster_configs_lambda
         self._allowed_cluster_config_overrides = allowed_cluster_config_overrides
         self._description = description
 
@@ -67,7 +74,7 @@ class EMRLaunchFunction(core.Construct):
         # Create Task for overriding cluster configurations
         override_cluster_configs = emr_tasks.OverrideClusterConfigs(
             self, 'OverrideClusterConfigsChain',
-            cluster_config=cluster_config.config,
+            cluster_config=cluster_configuration.config,
             override_cluster_configs_lambda=override_cluster_configs_lambda,
             allowed_cluster_config_overrides=allowed_cluster_config_overrides).task
         # Attach an error catch to the Task
@@ -113,21 +120,87 @@ class EMRLaunchFunction(core.Construct):
 
         self._ssm_parameter = ssm.StringParameter(
             self, 'SSMParameter',
-            string_value=json.dumps({
-                'LaunchFunctionName': launch_function_name,
-                'ClusterConfiguration': f'{cluster_config.namespace}/{cluster_config.configuration_name}',
-                'DefaultFailIfClusterRunning': default_fail_if_cluster_running,
-                'SuccessTopic': success_topic.topic_arn if success_topic is not None else None,
-                'FailureTopic': failure_topic.topic_arn if failure_topic is not None else None,
-                'OverrideClusterConfigsLambda':
-                    override_cluster_configs_lambda.function_arn
-                    if override_cluster_configs_lambda is not None
-                    else None,
-                'AllowedClusterConfigOverrides': self._allowed_cluster_config_overrides,
-                'StateMachine': self._state_machine.state_machine_arn,
-                'Description': self._description
-            }),
+            string_value=self._property_values_to_json(),
             parameter_name=f'{SSM_PARAMETER_PREFIX}/{namespace}/{launch_function_name}')
+
+    def _property_values_to_json(self):
+        return json.dumps({
+            'LaunchFunctionName': self._launch_function_name,
+            'Namespace': self._namespace,
+            'ClusterConfiguration': f'{self._cluster_configuration.namespace}/{self._cluster_configuration.configuration_name}',
+            'DefaultFailIfClusterRunning': self._default_fail_if_cluster_running,
+            'SuccessTopic': self._success_topic.topic_arn if self._success_topic is not None else None,
+            'FailureTopic': self._failure_topic.topic_arn if self._failure_topic is not None else None,
+            'OverrideClusterConfigsLambda':
+                self._override_cluster_configs_lambda.function_arn
+                if self._override_cluster_configs_lambda is not None
+                else None,
+            'AllowedClusterConfigOverrides': self._allowed_cluster_config_overrides,
+            'StateMachine': self._state_machine.state_machine_arn,
+            'Description': self._description
+        })
+
+    def _property_values_from_json(self, property_values):
+        self._launch_function_name = property_values['LaunchFunctionName']
+        self._namespace = property_values['Namespace']
+
+        config_parts = property_values['ClusterConfiguration'].split('/')
+        self._cluster_configuration = ClusterConfiguration.from_stored_configuration(
+            self, 'ClusterConfiguration', config_parts[1], config_parts[0])
+
+        self._default_fail_if_cluster_running = property_values['DefaultFailIfClusterRunning']
+
+        topic = property_values.get('SuccessTopic', None)
+        self._success_topic = sns.Topic.from_topic_arn(self, 'SuccessTopic', topic) \
+            if topic is not None \
+            else None
+
+        topic = property_values.get('FailureTopic', None)
+        self._failure_topic = sns.Topic.from_topic_arn(self, 'FailureTopic', topic) \
+            if topic is not None \
+            else None
+
+        func = property_values.get('OverrideClusterConfigsLambda', None)
+        self._override_cluster_configs_lambda = aws_lambda.Function.from_function_arn(
+            self, 'OverrideClusterConfigsLambda', func) \
+            if func is not None \
+            else None
+
+        self._allowed_cluster_config_overrides = property_values.get('AllowedClusterConfigOverrides', None)
+        self._description = property_values.get('Description', None)
+
+        state_machine = property_values['StateMachine']
+        self._state_machine = sfn.StateMachine.from_state_machine_arn(self, 'StateMachine', state_machine)
+
+        return self
+
+    @property
+    def launch_function_name(self) -> str:
+        return self._launch_function_name
+
+    @property
+    def namespace(self) -> str:
+        return self._namespace
+
+    @property
+    def cluster_configuration(self) -> ClusterConfiguration:
+        return self._cluster_configuration
+
+    @property
+    def default_fail_if_cluster_running(self) -> bool:
+        return self._default_fail_if_cluster_running
+
+    @property
+    def success_topic(self) -> sns.Topic:
+        return self._success_topic
+
+    @property
+    def failure_topic(self) -> sns.Topic:
+        return self._failure_topic
+
+    @property
+    def override_cluster_configs_lambda(self) -> aws_lambda.Function:
+        return self._override_cluster_configs_lambda
 
     @property
     def allowed_cluster_config_overrides(self) -> Mapping[str, str]:
@@ -166,3 +239,9 @@ class EMRLaunchFunction(core.Construct):
         except ClientError as e:
             if e.response['Error']['Code'] == 'ParameterNotFound':
                 raise EMRLaunchFunctionNotFoundError()
+
+    @staticmethod
+    def from_stored_function(scope: core.Construct, id: str, launch_function_name: str, namespace: str = 'default'):
+        stored_function = EMRLaunchFunction.get_function(launch_function_name, namespace)
+        launch_function = EMRLaunchFunction(scope, id, launch_function_name=launch_function_name, namespace=namespace)
+        return launch_function._property_values_from_json(stored_function)